### 2024-11-22 #7513 @editorialbot
[PRE REVIEW]: AdvSecureNet: A Python Toolkit for Adversarial Machine Learning
**Submitting author:** @melihcatal (<a href="http://orcid.org/0009-0009-0231-287X">Melih Catal</a>)
**Repository:** https://github.com/melihcatal/advsecurenet
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** V0.2.3
