### 2024-11-28 #7513 @osorensen
@melihcatal, I am looking for reviewers for your submission now. In the meantime, can you please fix the DOI issues for the references mentioned in the [comment above](https://github.com/openjournals/joss-reviews/issues/7513#issuecomment-2506605138)?
