### 2024-11-28 #7513 @osorensen
👋 @diningphil, @So-Cool, @asmitamitra, @lgiannantoni, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
