### 2024-11-29 6996b4a
Add missing DOIs to references
