+++ joss_paper/paper.bib
@@ -4,21 +4,24 @@
   journal = {CoRR},
   volume  = {1807.01069},
   year    = {2018},
-  url     = {https://arxiv.org/pdf/1807.01069}
+  url     = {https://arxiv.org/pdf/1807.01069},
+  doi     = {10.48550/arXiv.1807.01069}
 }
 
 @article{ding2019advertorch,
   title   = {{AdverTorch} v0.1: An Adversarial Robustness Toolbox based on PyTorch},
   author  = {Ding, Gavin Weiguang and Wang, Luyu and Jin, Xiaomeng},
   journal = {arXiv preprint arXiv:1902.07623},
-  year    = {2019}
+  year    = {2019},
+  doi     = {10.48550/arXiv.1902.07623}
 }
 
 @article{melis2019secml,
   title   = {secml: A Python Library for Secure and Explainable Machine Learning},
   author  = {Melis, Marco and Demontis, Ambra and Pintor, Maura and Sotgiu, Angelo and Biggio, Battista},
   journal = {arXiv preprint arXiv:1912.10013},
-  year    = {2019}
+  year    = {2019},
+  doi     = {10.1016/j.softx.2022.101095}
 }
 
 
@@ -40,7 +43,8 @@
   author    = {Dong, Yinpeng and Fu, Qi-An and Yang, Xiao and Pang, Tianyu and Su, Hang and Xiao, Zihao and Zhu, Jun},
   booktitle = {Proceedings of the IEEE/CVF Conference on Computer Vision and Pattern Recognition (CVPR)},
   pages     = {321--331},
-  year      = {2020}
+  year      = {2020},
+  doi       = {10.1109/CVPR42600.2020.00040}
 }
 
 
@@ -50,7 +54,8 @@
   year          = {2018},
   eprint        = {1610.00768},
   archiveprefix = {arXiv},
-  primaryclass  = {cs.LG}
+  primaryclass  = {cs.LG},
+  doi           = {10.48550/arXiv.1610.00768}
 }
 
 
@@ -62,14 +67,16 @@
   number    = {1},
   pages     = {84},
   year      = {2013},
-  publisher = {Technopark Publications}
+  publisher = {Technopark Publications},
+  doi       = {10.48550/arXiv.1403.0485}
 }
 
 @article{bojarski2016end,
   title   = {End to end learning for self-driving cars},
   author  = {Bojarski, Mariusz and Del Testa, Davide and Dworakowski, Daniel and Firner, Bernhard and Flepp, Beat and Goyal, Prasoon and Jackel, Lawrence D and Monfort, Mathew and Muller, Urs and Zhang, Jiakai and others},
   journal = {arXiv preprint arXiv:1604.07316},
-  year    = {2016}
+  year    = {2016},
+  doi     = {10.48550/arXiv.1604.07316}
 }
 
 
@@ -82,7 +89,8 @@
   number    = {5},
   pages     = {868--869},
   year      = {2023},
-  publisher = {Springer}
+  publisher = {Springer},
+  doi       = {10.1007/s10439-023-03172-7}
 }
 
 @article{mintz2019introduction,
@@ -93,7 +101,8 @@
   number    = {2},
   pages     = {73--81},
   year      = {2019},
-  publisher = {Taylor \& Francis}
+  publisher = {Taylor \& Francis},
+  doi       = {10.1080/13645706.2019.1575882}
 }
 
 
@@ -102,7 +111,8 @@
   author    = {Ian Goodfellow and Jonathon Shlens and Christian Szegedy},
   year      = {2015},
   url       = {http://arxiv.org/abs/1412.6572},
-  booktitle = {International Conference on Learning Representations}
[…]
