### 2024-11-29 #7513 @melihcatal
> @melihcatal, I am looking for reviewers for your submission now. In the meantime, can you please fix the DOI issues for the references mentioned in the [comment above](https://github.com/openjournals/joss-reviews/issues/7513#issuecomment-2506605138)?

Thank you for starting the process of finding reviewers! I’ve updated the references, and the DOI issue seems to be resolved now.
