### 2024-12-02 #7513 @lgiannantoni
Hi Øystein,

Thanks for the invitation.
I am available to review the article.

-- Leonardo

Il gio 28 nov 2024, 19:10 Øystein Sørensen ***@***.***> ha
scritto:

> 👋 @diningphil <https://github.com/diningphil>, @So-Cool
> <https://github.com/So-Cool>, @asmitamitra
> <https://github.com/asmitamitra>, @lgiannantoni
> <https://github.com/lgiannantoni>, would any of you be willing to review
> this submission for JOSS? We carry out our checklist-driven reviews here in
> GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
