### 2024-12-16 melihcatal/advsecurenet#148 @So-Cool
JOSS review comments
Below you can find my JOSS review comments (openjournals/joss-reviews#7576).
