### 2024-12-16 melihcatal/advsecurenet#148 @So-Cool
# Paper

- it may be worth sorting the in-text citations by year

- space missing before the `[@biswas2023role; @mintz2019introduction]` citation
https://github.com/melihcatal/advsecurenet/blob/6996b4a4cbd32980c21f07526082fdef23cee9bc/joss_paper/paper.md?plain=1#L38

- how is "natural language processing" a "critical application"?
https://github.com/melihcatal/advsecurenet/blob/6996b4a4cbd32980c21f07526082fdef23cee9bc/joss_paper/paper.md?plain=1#L38

- the rest of the list is in present tense, but "Optimized" is in past tense
https://github.com/melihcatal/advsecurenet/blob/6996b4a4cbd32980c21f07526082fdef23cee9bc/joss_paper/paper.md?plain=1#L48

- same here, should probably be "integrates seamlessly" rather than "seamless integration"
https://github.com/melihcatal/advsecurenet/blob/6996b4a4cbd32980c21f07526082fdef23cee9bc/joss_paper/paper.md?plain=1#L53

- both tables need a bit of restructuring as they are otherwise difficult to read
    * the *Dataset* and *Single GPU Time (min)* in the second table are too close, hence the column entries are confusing

- should this sentence just refer to features and not performance...
https://github.com/melihcatal/advsecurenet/blob/6996b4a4cbd32980c21f07526082fdef23cee9bc/joss_paper/paper.md?plain=1#L55
...since the sentence before the next table mentions performance (also is should probably end with full stop for consistency)
https://github.com/melihcatal/advsecurenet/blob/6996b4a4cbd32980c21f07526082fdef23cee9bc/joss_paper/paper.md?plain=1#L73

# References

- Biswas et al: "chat gpt" should probably be "Chat GPT"
- Melis et al is an incorrect citation – it refers to arXiv but the DOI points to a journal article; also the given title is different to that of the journal paper; in the title "Secml" should be `secml`, "python" should be "Python", etc.
- Parmar et all is cited as "International Journal of Computer Technology and Applications" but has arXiv DOI
    * same issue for Khalid et al
- "nlp" in Patwardhan et al should be capitalised
- Szegedy et al has "CoRR, abs/1312.6199" as journal, but other arXiv citations do not
    * same issue for Nicolae et al

# Motivation

- since IBM Art appears to have a much more comprehensive set of features, why do you think that developing a new package is a better approach than contributing multi-GPU support as well as CLI and config file functionality to IBM Art?
