### 2024-12-16 melihcatal/advsecurenet#148 @So-Cool
# Usage

- generating a config file appears not to work as per readme
https://github.com/melihcatal/advsecurenet?tab=readme-ov-file#command-line-tool
```
$ advsecurenet utils configs get -c fgsm_attack_base_config.yml -o fgsm.yml

Please provide either the --save or --print-output flag to save or print the configuration file!
```

## Examples

- in this example the last cell should use the `default_fgsm_attack_config.yml` file
https://github.com/melihcatal/advsecurenet/blob/main/examples/cli/adversarial_attacks/distributed/fgsm/fgsm.ipynb

- same issue here and probably in other examples
https://github.com/melihcatal/advsecurenet/blob/main/examples/cli/adversarial_attacks/distributed/lots/lots.ipynb

- also running this example fails

```
$ advsecurenet attack lots -c default_lots_attack_config.yml

Files already downloaded and verified
                                                                                                                                              Traceback (most recent call last):
  File "bin/advsecurenet", line 8, in <module>
    sys.exit(main())
             ^^^^^^
  File "lib/python3.12/site-packages/click/core.py", line 1157, in __call__
    return self.main(*args, **kwargs)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "lib/python3.12/site-packages/click/core.py", line 1078, in main
    rv = self.invoke(ctx)
         ^^^^^^^^^^^^^^^^
  File "lib/python3.12/site-packages/click/core.py", line 1688, in invoke
    return _process_result(sub_ctx.command.invoke(sub_ctx))
                           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "lib/python3.12/site-packages/click/core.py", line 1688, in invoke
    return _process_result(sub_ctx.command.invoke(sub_ctx))
                           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "lib/python3.12/site-packages/click/core.py", line 1434, in invoke
    return ctx.invoke(self.callback, **ctx.params)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "lib/python3.12/site-packages/click/core.py", line 783, in invoke
    return __callback(*args, **kwargs)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "lib/python3.12/site-packages/cli/commands/attack/commands.py", line 543, in lots
    cli_attack("LOTS", config, **kwargs)
  File "lib/python3.12/site-packages/cli/logic/attack/attack.py", line 46, in cli_attack
    raise e
  File "lib/python3.12/site-packages/cli/logic/attack/attack.py", line 42, in cli_attack
    attacker.execute()
  File "lib/python3.12/site-packages/cli/logic/attack/attacker.py", line 54, in execute
    self._execute_attack()
  File "lib/python3.12/site-packages/cli/logic/attack/attacker.py", line 65, in _execute_attack
    adv_imgs = attacker.execute()
               ^^^^^^^^^^^^^^^^^^
  File "lib/python3.12/site-packages/advsecurenet/attacks/attacker/attacker.py", line 33, in execute
    return self._execute_attack()
           ^^^^^^^^^^^^^^^^^^^^^^
  File "lib/python3.12/site-packages/advsecurenet/attacks/attacker/attacker.py", line 89, in _execute_attack
    adv_images = self._generate_adversarial_images(
                 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "lib/python3.12/site-packages/advsecurenet/attacks/attacker/attacker.py", line 143, in _generate_adversarial_images
    return self._config.attack.attack(self._model, images, labels, target_images)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "lib/python3.12/site-packages/advsecurenet/attacks/gradient_based/lots.py", line 76, in attack
    x, x_target = self._prepare_inputs(x, x_target)
                  ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "lib/python3.12/site-packages/advsecurenet/attacks/gradient_based/lots.py", line 296, in _prepare_inputs
    x_target = x_target.clone().detach().requires_grad_(False)
               ^^^^^^^^^^^^^^
AttributeError: 'NoneType' object has no attribute 'clone'

```

- do you need to download the entire directory to run an example, or are the commands suppose to generate the config file from scratch?

# Repository

- CI – Quality Assurance appears to have failed
https://github.com/melihcatal/advsecurenet/actions/runs/11628011909

- redundant files in the repository
    * `advsecurenet/docs/.DS_Store`

# Documentation

- no examples that would make getting started with the package easier
    * the *examples* URL on this page <https://melihcatal.github.io/advsecurenet/introduction.html#examples> does not work
    https://melihcatal.github.io/advsecurenet/examples/

- there should probably be a separate *getting started* page in the documentation to introduce the users to a couple of simple examples
    * having the examples as a standalone directory in your repository is probably insufficient
    https://github.com/melihcatal/advsecurenet/tree/main/examples
