### 2024-12-16 #7576 @melihcatal
> @melihcatal, I've left my comments in [melihcatal/advsecurenet#148](https://github.com/melihcatal/advsecurenet/issues/148). I'll update my review checklist once you address the outstanding issues.

Thank you very much for your reviews and for starting the review round. I will address your reviews as soon as possible and will keep you updated. 🙂
