+++ _sources/Introduction.rst.txt
@@ -1,416 +1,0 @@
-AdvSecureNet
-============
-
-Build and CI/CD Status
-~~~~~~~~~~~~~~~~~~~~~~~~~
-|Unit Tests and Style
-Checks| |Build and Deploy Sphinx Documentation| |Upload Python Package|
-
-Code Quality and Coverage
-~~~~~~~~~~~~~~~~~~~~~~~~~
-|Quality Gate Status| |Bugs| |Code Smells| |Coverage| |Duplicated Lines
-(%)| |Lines of Code| |Reliability Rating| |Security Rating| |Technical
-Debt| |Maintainability Rating| |Vulnerabilities| 
-
-Package Information
-~~~~~~~~~~~~~~~~~~~~~~~~~
-|Code Style: Black| |pypi| |gitflow|
-
-
-.. raw:: html
-
-   <p align="center">
-
-.. raw:: html
-
-   </p>
-
-AdvSecureNet is a Python library for Machine Learning Security,
-developed by `Melih Catal <https://github.com/melihcatal>`__ at
-`University of Zurich <https://www.uzh.ch/en.html>`__ as part of his
-Master’s Thesis under the supervision of `Prof. Dr. Manuel
-Günther <https://www.ifi.uzh.ch/en/aiml/people/guenther.html>`__. The
-main focus of the library is on adversarial attacks and defenses for
-vision tasks, with plans to extend support to other tasks such as
-natural language processing.
-
-The library provides tools to generate adversarial examples, evaluate
-the robustness of machine learning models against adversarial attacks,
-and train robust machine learning models. Built on top of
-`PyTorch <https://pytorch.org/>`__, it is designed to be modular and
-extensible, making it easy to run experiments with different
-configurations. AdvSecureNet supports multi-GPU setups to enhance
-computational efficiency and fully supports both CLI and API interfaces,
-along with external YAML configuration files, enabling comprehensive
-testing and evaluation, facilitating the sharing and reproducibility of
-experiments.
-
-
-Features
---------
-
-**Adversarial Attacks:** AdvSecureNet supports a diverse range of
-evasion attacks on computer vision tasks, including gradient-based,
-decision-based, single-step, iterative, white-box, black-box, targeted,
-and untargeted attacks, enabling comprehensive testing and evaluation of
-neural network robustness against various types of adversarial examples.
-
-**Adversarial Defenses:** The toolkit includes adversarial training and
-ensemble adversarial training. Adversarial training incorporates
-adversarial examples into the training process to improve model
-robustness, while ensemble adversarial training uses multiple models or
-attacks for a more resilient defense strategy.
-
-**Evaluation Metrics:** AdvSecureNet supports metrics like accuracy,
-robustness, transferability, and similarity. Accuracy measures
-performance on clean data, robustness assesses resistance to attacks,
-transferability evaluates how well adversarial examples deceive
-different models, and similarity quantifies perceptual differences using
-PSNR and SSIM.
-
-**Multi-GPU Support:** AdvSecureNet is optimized for multi-GPU setups,
-enhancing the efficiency of training, evaluation, and adversarial attack
-generation, especially for large models and datasets or complex methods.
-By utilizing multiple GPUs in parallel, AdvSecureNet aims to reduce
-computational time, making it ideal for large-scale experiments and deep
-learning models.
-
-**CLI and API Interfaces:** AdvSecureNet offers both CLI and API
-interfaces. The CLI allows for quick execution of attacks, defenses, and
-evaluations, while the API provides advanced integration and extension
-within user applications.
-
-**External Configuration Files:** The toolkit supports YAML
[…]
